Update for py3 and wxWidgets 4.2.0
The main changes here are to update chronolapse for python3 (tested with
python 3.11.0) and wxWidgets 4.2.0.

Additionally, there's a small bugfix to the quoting of the output file
for the menconcoder command (allows the output file path to contain
spaces), and adds a new set of config variables to allow explicitly
setting the resolution of the webcam capture.